fix: prevent types issues w/ conditional tabs
